+++ timezones/constants.py
@@ -2,13 +2,6 @@
 Constants
 """
 
-# Standard Library
-import os
-
-PACKAGE_NAME = "timezones"
-APP_BASE_DIR = os.path.join(os.path.dirname(__file__))
-APP_STATIC_DIR = os.path.join(APP_BASE_DIR, "static", PACKAGE_NAME)
-
 AA_TIMEZONE_DEFAULT_PANELS: list[dict[str, str | dict[str, str]]] = [
     {
         "panel_name": "US / Pacific",