+++ CHANGELOG.md
@@ -35,6 +35,10 @@ Section Order:
 ### Security
 -->
 
+### Removed
+
+- Unused constants
+
 ## [2.4.0] - 2025-06-03
 
 ### Changed